Refactor and added News Loader and Band Loader
Did a lot of refactoring to make the project more manageable, Added a
news loaded with styling, re-styled the upcoming gigs so we can fit more
in the area. Moved gigs to the middle as to draw attention, styled hover
interactions as well as adding a bands loader to replace the gallery.
This is so people who visit the site can go directly to the band pages
when they are made.